+++ test/analyzers/lint_analyzer/rules/rules_list/binary_expression_operand_order/binary_expression_operand_order_test.dart
@@ -1,6 +1,6 @@
 @TestOn('vm')
-import 'package:dart_code_metrics/src/analyzers/lint_analyzer/rules/rules_list/binary_expression_operand_order/binary_expression_operand_order.dart';
 import 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/severity.dart';
+import 'package:dart_code_metrics/src/analyzers/lint_analyzer/rules/rules_list/binary_expression_operand_order/binary_expression_operand_order.dart';
 import 'package:test/test.dart';
 
 import '../../../../../helpers/rule_test_helper.dart';