+++ test/analyzers/lint_analyzer/rules/rules_list/avoid_unused_parameters/examples/incorrect_example.dart
@@ -1,7 +1,6 @@
 class TestClass implements AbstractClass {
   String value;
 
-  // LINT
   @override
   void abstractMethod(String string) {}