+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Add static code diagnostics `always-remove-listener`, `avoid-wrapping-in-padding` and `avoid-unnecessary-setstate`.
-* Improve static code diagnostic `avoid-returning-widgets`.
+* Improve static code diagnostics `avoid-returning-widgets` and `avoid-unused-parameters`.
 * Remove deprecated `Lines of Executable Code` metric, use `Source lines of Code` instead.
 * Changed the supported `analyzer` version to `^1.7.0`.
 * Introduce `analyze` and `check-unused-files` commands.